+++ lib/animated_emoji.dart
@@ -1,11 +1,13 @@
 /// <picture>
 ///   <source srcset="https://fonts.gstatic.com/s/e/notoemoji/latest/1f680/512.avif" type="image/avif">
 ///   <source srcset="https://fonts.gstatic.com/s/e/notoemoji/latest/1f680/512.webp" type="image/webp">
-///   <img src="https://fonts.gstatic.com/s/e/notoemoji/latest/1f680/512.gif" alt="🚀" width="32" height="32">
+///   <img src="https://fonts.gstatic.com/s/e/notoemoji/latest/1f680/512.gif" alt="🚀" width="64" height="64">
 /// </picture>
 ///
 /// Library for using `AnimatedEmoji`.
 library animated_emoji;
 
 export 'emoji.dart';
-export 'emojis.dart';
+export 'emoji_data.dart';
+export 'emojis.g.dart';
+export 'util.dart';